+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 + Added "Bow Aim" module (not finished so no code is inside, just the base module)
 + Added thicker window icons
 + Edits to "Glide" option in Long Jump
++ Added options "FPS", "TPS", "Potions" to Interface module
 ```
 
 ---